+++ src/frontend/exports.js
@@ -1,0 +1,27 @@
+// components
+export { Card } from "./components/Card.jsx";
+export { Nav } from "./components/Nav";
+
+// images
+import logo from "./assets/logo/logo.png";
+import coverImage from "./assets/images/cover.jpeg";
+import bestSellers from "./assets/images/bestsellers.jpeg";
+import newReleases from "./assets/images/new release.jpeg";
+import expertPicks from "./assets/images/expert.jpeg";
+import contemporaryFiction from "./assets/images/book cluster.jpeg";
+import selfHelp from "./assets/images/self help.jpeg";
+import biographies from "./assets/images/bio.jpeg";
+import spirituality from "./assets/images/spiritual.jpeg";
+import mythology from "./assets/images/history.jpeg";
+export {
+  logo,
+  coverImage,
+  bestSellers,
+  newReleases,
+  expertPicks,
+  contemporaryFiction,
+  selfHelp,
+  biographies,
+  spirituality,
+  mythology,
+};
